# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ecommerce API in go using Gin frame work",
        "title": "E Commerce API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/user": {
            "get": {
                "description": "get list of all registered admins",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Users"
                ],
                "summary": "list of users",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.UpdateUserData"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch user data",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/admin/user/{id}": {
            "delete": {
                "description": "Delete an existing user from admin side",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Users"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "500": {
                        "description": "Failed to delete user",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            },
            "patch": {
                "description": "Edit user details based on user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Users"
                ],
                "summary": "Edit user details",
                "parameters": [
                    {
                        "description": "Update user info",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.UpdateUserData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "406": {
                        "description": "Failed to bind data",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "500": {
                        "description": "Failed to update details",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/product/rating/{id}": {
            "post": {
                "description": "Product rating store",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Rating store",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Failed to create rating",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "description": "Get a paginated list of products including product name, description, stock, price, brand name, and image.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "product details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "404": {
                        "description": "Product details not found",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/user/forgotpass": {
            "post": {
                "description": "Sends an OTP to the user's email for password recovery if the email exists in the database.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Send OTP for password recovery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's email address",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/user/forgotpass/otp": {
            "post": {
                "description": "Checks if the provided OTP is valid and not expired for password recovery.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Check OTP validity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's OTP",
                        "name": "otp",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/user/new-password": {
            "patch": {
                "description": "Sets a new password for the user after verifying the OTP.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Set new password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "New password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.UpdateUserData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}